chore: rename package ami to awsami
