+++ pkg/core/pipeline/source/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/updatecli/updatecli/pkg/core/pipeline/scm"
 	"github.com/updatecli/updatecli/pkg/core/result"
 	"github.com/updatecli/updatecli/pkg/core/transformer"
-	"github.com/updatecli/updatecli/pkg/plugins/aws/ami"
+	"github.com/updatecli/updatecli/pkg/plugins/awsami"
 	"github.com/updatecli/updatecli/pkg/plugins/docker"
 	"github.com/updatecli/updatecli/pkg/plugins/file"
 	gitTag "github.com/updatecli/updatecli/pkg/plugins/git/tag"
@@ -155,7 +155,7 @@ func (s *Source) Run() (err error) {
 func (s *Source) Unmarshal() (sourcer Sourcer, changelog Changelog, err error) {
 	switch s.Config.Kind {
 	case "aws/ami":
-		a := ami.AMI{}
+		a := awsami.AMI{}
 
 		err := mapstructure.Decode(s.Config.Spec, &a.Spec)
