+++ pkg/plugins/awsami/condition.go
@@ -1,4 +1,4 @@
-package ami
+package awsami
 
 import (
 	"errors"
