+++ pkg/plugins/awsami/filter.go
@@ -1,4 +1,4 @@
-package ami
+package awsami
 
 import "fmt"
