+++ pkg/plugins/awsami/helpers.go
@@ -1,4 +1,4 @@
-package ami
+package awsami
 
 import (
 	"fmt"
