+++ pkg/plugins/awsami/source.go
@@ -1,4 +1,4 @@
-package ami
+package awsami
 
 import (
 	"strings"
